requirements.txt
python-dotenv==1.1.1
openai==1.93.0
pandas==2.3.2
numpy==1.24.3
scipy==1.11.4
scikit-learn==1.3.2
xgboost==2.0.2
lightgbm==4.1.0
matplotlib==3.8.2
seaborn==0.13.0
e2b-code-interpreter
google-cloud-bigquery==3.21.0
db-dtypes==1.4.3
imbalanced-learn==0.11.0
shap==0.43.0
optuna==3.4.0